// UI/src/apiclient/mockedItemApiClient.ts
export class MockedItem implements IMockedItem {
    code?: string | undefined;
    description?: string | undefined;
    businessLine?: string | undefined;
    itemType?: string | undefined;
    solution?: string | undefined;
    product?: string | undefined;
    cisco?: string | undefined;
    saleArticalGroup?: string | undefined;
    stockBtwCode?: string | undefined;

    constructor(data?: IMockedItem) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.code = _data["code"];
            this.description = _data["description"];
            this.businessLine = _data["businessLine"];
            this.itemType = _data["itemType"];
            this.solution = _data["solution"];
            this.product = _data["product"];
            this.cisco = _data["cisco"];
            this.saleArticalGroup = _data["saleArticalGroup"];
            this.stockBtwCode = _data["stockBtwCode"];
        }
    }

    static fromJS(data: any): MockedItem {
        data = typeof data === 'object' ? data : {};
        let result = new MockedItem();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["code"] = this.code;
        data["description"] = this.description;
        data["businessLine"] = this.businessLine;
        data["itemType"] = this.itemType;
        data["solution"] = this.solution;
        data["product"] = this.product;
        data["cisco"] = this.cisco;
        data["saleArticalGroup"] = this.saleArticalGroup;
        data["stockBtwCode"] = this.stockBtwCode;
        return data;
    }
}

export interface IMockedItem {
    code?: string | undefined;
    description?: string | undefined;
    businessLine?: string | undefined;
    itemType?: string | undefined;
    solution?: string | undefined;
    product?: string | undefined;
    cisco?: string | undefined;
    saleArticalGroup?: string | undefined;
    stockBtwCode?: string | undefined;
}